+++ Backend/Services/SaleService/Dtos/InvoiceReadDto.cs
@@ -10,7 +10,7 @@ namespace SaleService.Dtos
         public int CustomerId { get; set; }
         public int? PromotionId { get; set; }
         public string? PaymentMethod { get; set; }
-        public int? Total { get; set; }
+        public float? Total { get; set; }
         public string? Status { get; set; }
         public IEnumerable<Schedule> Schedules { get; set; }
         public Employee Employee { get; set; }
